Splits up configuration into 3 files.
